Resolved a bugg where TempData would not properly proc

+++ InventarVali/Areas/Admin/Controllers/ComputerController.cs
@@ -136,15 +136,20 @@ namespace InventarVali.Areas.Admin.Controllers
                 return Json(new { success = false, message = "Error while deleting" });
             }
 
-            var oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, computerToBeDeleted.ImageUrl.TrimStart('\\'));
-
-            if (System.IO.File.Exists(oldImagePath))
+            if (!string.IsNullOrEmpty(computerToBeDeleted.ImageUrl))
             {
-                System.IO.File.Delete(oldImagePath);
+                var oldImagePath = Path.Combine(_webHostEnvironment.WebRootPath, computerToBeDeleted.ImageUrl.TrimStart('\\'));
+
+                if (System.IO.File.Exists(oldImagePath))
+                {
+                    System.IO.File.Delete(oldImagePath);
+                }
+
             }
+            
             _unitOfWork.Computer.Remove(computerToBeDeleted);
             _unitOfWork.Save();
-
+            TempData["success"] = "Computer was deleted successfully";
             return Json(new {success = true, message = "Deleted Successfully"});
         }
         #endregion
